:lock: Improve key derivation by using argon2
